requirements.txt
line-bot-sdk
flask
google-generativeai
Pillow
googlemaps
numpy
beautifulsoup4
requests
googlesearch-python